combine util.h&utilpp.h
