+++ selfdrive/common/params.cc
@@ -18,8 +18,6 @@
 #include <string.h>
 
 #include "common/util.h"
-#include "common/utilpp.h"
-
 
 #if defined(QCOM) || defined(QCOM2)
 const std::string default_params_path = "/data/params";
