+++ selfdrive/loggerd/loggerd.cc
@@ -31,7 +31,6 @@
 #include "common/params.h"
 #include "common/swaglog.h"
 #include "common/visionipc.h"
-#include "common/utilpp.h"
 #include "common/util.h"
 #include "camerad/cameras/camera_common.h"
 #include "logger.h"
