+++ selfdrive/ui/ui.cc
@@ -12,7 +12,6 @@
 #include "common/util.h"
 #include "common/swaglog.h"
 #include "common/visionimg.h"
-#include "common/utilpp.h"
 #include "ui.hpp"
 #include "paint.hpp"
